% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfConn.R
\docType{data}
\name{fMRI_dataList_shrinked}
\alias{fMRI_dataList_shrinked}
\title{A shrinked sample subject functional magnetic resonance imaging (fMRI) time series data}
\format{A list which contains a matrix with 2 rows and 19 columns:
\describe{
  \item{row}{index of regions of interest (ROIs).}
  \item{column}{fMRI time series data at each time point.}
}}
\usage{
data(fMRI_dataList_shrinked)
}
\description{
Contains a list of data matrices with consisting time series data of only one subject, and only 1 scan is preserved. This shrinked dataset is useful for user to test the programs and functions.
}
\keyword{datasets}
