% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfConn.R
\docType{data}
\name{MLPB_output_median}
\alias{MLPB_output_median}
\title{Sample median output of Multivariate Linear Process Bootstrap (MLPB)}
\format{A list with 3 matrices, each matrix has 5 rows and 731 columns, where rows represent subjects-level multivariate linear process bootstrapping (MLPB) estimates and columns represent median of MLPB estimates at each window.
\describe{

  \item{median_1_2}{matrix with 5 rows and 731 columns, containing median of multivariate linear process bootstrapping results between region 1 and 2 in subject-level.}
  \item{median_1_3}{matrix with 5 rows and 731 columns, containing median of multivariate linear process bootstrapping results between region 1 and 3 in subject-level.}
  \item{median_2_3}{matrix with 5 rows and 731 columns, containing median of multivariate linear process bootstrapping results between region 2 and 3 in subject-level.}
}}
\usage{
data(MLPB_output_median)
}
\description{
A dataset containing summarized by calculating the median for dynamic functional connectivity estimates of pairwise region of interest (ROI) comparisions.
}
\keyword{datasets}
